requirements.txt
pettingzoo==1.24.0
agilerl==0.1.21
torch==2.1.0
pygame==2.5.2
wandb==0.13.10
python-dotenv==1.0.1